+++ hypermedia/tornado/__init__.py
@@ -1,0 +1,9 @@
+"""
+Tornado companion library.
+
+This module contains companion classes and utility functions that
+are designed to be used with the `Tornado`_ web framework.
+
+.. _Tornado: http://tornadoweb.org/
+
+"""