+++ tests/compat.py
@@ -1,0 +1,9 @@
+import unittest
+
+try:
+    from unittest import mock
+except ImportError:
+    # noinspection PyPackageRequirements
+    import mock
+
+__all__ = ('mock', 'unittest')